requirements.txt
python-digitalocean==1.17.0
requests==2.31.0
tabulate==0.9.0
